Moved preloder includes, checked for ipc vulnerabilities (#1377)
* moved preloder includes, checked popupwindows

* small spell change
